add check_boxes input
